adjusted create to have right data types

+++ HW_1/create_scripts.sql
@@ -1,12 +1,13 @@
 use twitter_4300;
 CREATE TABLE IF NOT EXISTS twitter_4300.tweets (
-	`tweet_id` long,
-    `user_id` long,
+	`tweet_id` int(8),
+    `user_id` int(8),
     `tweet_ts` datetime,
     `tweet_text` text
     )
     ;
 CREATE TABLE IF NOT EXISTS twitter_4300.followers (
-	user_id long,
-    follows_id long
+	user_id int(8),
+    follows_id int(8)
     );
+
